\name{osm-router-package}
\alias{osm-router-package}
\alias{osm-router}
\docType{package}
\title{
\packageTitle{osm-router}
}
\description{
\packageDescription{osm-router}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{osm-router}
\packageIndices{osm-router}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{osm-router}

Maintainer: \packageMaintainer{osm-router}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
